chore(org): move firebase functions
